Rearrange module organization
