+++ src/psd2svg/version.py
@@ -1,1 +1,1 @@
-__version__ = '0.1.0a3'
+__version__ = '0.1.0a4'
