+++ tox.ini
@@ -1,9 +1,9 @@
 
-; this is a tox config for running psd-tools tests
+; this is a tox config for running psd2svg tests
 ; under different Python interpreters
 
 [tox]
-envlist = py27,py36
+envlist = py27,py35,py36
 
 [testenv]
 deps=
